move: map route 폴더 정리 (#46)
