+++ src/components/product/product-info-container/product-info-container.tsx
@@ -1,4 +1,4 @@
-import TagContainer from '@/app/map/(components)/tag-container/tag-container';
+import TagContainer from '@/app/map/_components/tag-container/tag-container';
 import ProductInfoFooter from '@/components/product/product-info-container/(components)/footer';
 import { ProductInfoContainerProps } from '@/types/bag';
 import * as styles from './styles.css';
