+++ src/components/product/product-info-thumb-container/product-info-thumb-container.tsx
@@ -1,4 +1,4 @@
-import TagContainer from '@/app/map/(components)/tag-container/tag-container';
+import TagContainer from '@/app/map/_components/tag-container/tag-container';
 import ProductInfoFooter from '@/components/product/product-info-container/(components)/footer';
 import { BagInfoResponse } from '@/types/bag';
 import * as styles from './styles.css';
